feat: tournament thumbnails, carousel fix, owner/employee access model, mobile UX
- Tournament create/edit: cover thumbnail upload to tournament-thumbnails bucket
  (+ cover_path in admin-tournaments-create/update, card cover preview)
- Fix BannerSlideshow stuck on 'Loading carousel' (measure width on placeholder)
- Owner/Employee access: role switcher at login, super key verification,
  request-access flow for employees, owner approve/reject + auto key handoff
- Per-admin permission toggles (custom_permissions isolated per assignment)
- Edge functions: admin-request-access, admin-permissions-update/list,
  owner-admins-reject; owner-admins-list email enrichment
- Migrations 0020-0025: security hardening, policy links, admin status rejected,
  request flow, service_role grants, public RPC wrappers
- Service_role grants + public schema wrappers for app.* RPC functions
- Reconciliation return-type fix
- Dark theme polish, responsive modals (maxHeight), mobile-friendly grids

+++ supabase/migrations/0015_rpc_functions.sql
@@ -29,7 +29,7 @@ begin
       -- Simplified: recompute available as sum(case when direction='credit' then amount when direction='debit' then -amount else 0 end)
       -- plus hold/release handled as moves. For MVP: just compare available via sum of credit/debit where direction in ('credit','debit')
       -- Real prod should compute precisely matching app.wallet_* semantics.
-      sum(case when direction='credit' then amount when direction in ('debit') then -amount else 0 end) as ledger_avail
+      sum(case when direction='credit' then amount when direction in ('debit') then -amount else 0 end)::bigint as ledger_avail
     from app.wallet_ledger l
     group by l.profile_id
   )
